Authentication and authorization

+++ app/app.py
@@ -6,6 +6,8 @@ from contextlib import asynccontextmanager
 from sqlalchemy.future import select
 from app.images import imagekit
 import os 
+from app.users import auth_backend,fastapi_users,current_active_user
+from app.schemas import UserRead,UserCreate,UserUpdate
 
 @asynccontextmanager
 async def lifespan(app:FastAPI):
@@ -14,11 +16,36 @@ async def lifespan(app:FastAPI):
 
 app=FastAPI(lifespan=lifespan)
 
-
+app.include_router(
+    fastapi_users.get_auth_router(auth_backend),
+    prefix="/auth",
+    tags=["auth"],
+)
+app.include_router(
+    fastapi_users.get_register_router(UserRead,UserCreate),
+    prefix="/auth",
+    tags=["auth"],
+)
+app.include_router(
+    fastapi_users.get_reset_password_router(),
+    prefix="/auth",
+    tags=["auth"],
+)   
+app.include_router(
+    fastapi_users.get_verify_router(UserRead),  
+    prefix="/auth",
+    tags=["auth"],
+)
+app.include_router(
+    fastapi_users.get_users_router(UserRead,UserUpdate),    
+    prefix="/users",
+    tags=["users"],
+)
 
 @app.post("/upload")
 async def upload_file(file:UploadFile=File(...),
                       caption:str=Form(""),
+                      user:UserRead=Depends(current_active_user),
                       session: AsyncSession=Depends(get_async_session)):
     try:
         file_bytes = await file.read()
@@ -32,6 +59,7 @@ async def upload_file(file:UploadFile=File(...),
 
         post = Post(
             caption=caption,
+            user_id=user.id,
             file_id=upload_result.file_id,
             url=upload_result.url,
             file_type="video" if (file.content_type or "").startswith("video") else "image",
@@ -48,31 +76,39 @@ async def upload_file(file:UploadFile=File(...),
         await file.close()
 
 @app.get("/feed")
-async def get_feed(session: AsyncSession = Depends(get_async_session)):
+async def get_feed(session: AsyncSession = Depends(get_async_session),user:UserRead=Depends(current_active_user)):
     result = await session.execute(select(Post).order_by(Post.created_at.desc()))
     posts=[row[0] for row in result.all()]
+    users=[row[0] for row in result.all()]
+    user_dict={u.id:u.email for u in users}
 
     posts_data=[]
 
     for post in posts:
         posts_data.append({
             "id":str(post.id),
+            "user_id":str(post.user_id),
             "caption":post.caption,
             "url":post.url,
             "file_type":post.file_type,
             "file_name":post.file_name,
-            "created_at":post.created_at.isoformat()
+            "created_at":post.created_at.isoformat(),
+            "is_owner":post.user_id==user.id,
+            "email":user_dict.get(post.user_id,"Unknown") 
         })
     return posts_data
 
 @app.delete("/delete/{post_id}")
-async def delete_post(post_id:str,session:AsyncSession=Depends(get_async_session)): 
+async def delete_post(post_id:str,session:AsyncSession=Depends(get_async_session),user:UserRead=Depends(current_active_user)): 
     
     result=await session.execute(select(Post).where(Post.id==post_id))
     post=result.scalar_one_or_none()
 
     if not post:
         raise HTTPException(status_code=404,detail="Post not found")
+    
+    if post.user_id != user.id:
+        raise HTTPException(status_code=403,detail="You are not authorized to delete this post")
 
     try:
         imagekit.files.delete(file_id=post.file_id)
